Make room for pipeline status
